Add unit tests

+++ src/MyTTCBot/Models/UserContext.cs
@@ -9,9 +9,9 @@ namespace MyTTCBot.Models
 
     public class UserLocation
     {
-        public long Latitude { get; set; }
+        public double Latitude { get; set; }
 
-        public long Longitude { get; set; }
+        public double Longitude { get; set; }
 
         public DateTime UpdatedAt { get; set; }
     }
